mnCharSel_802640A0: hold the player count in a byte
It only ever holds 1, 2 or 4 and mirrors the u8 global it is written beside;
as an s32 its web outranks the compiler temps it shares the band with. 24 fewer
differing instructions.

+++ src/melee/mn/mncharsel.c
@@ -3957,7 +3957,7 @@ s32 mnCharSel_802640A0(void)
     HSD_JObj* jobj;
     HSD_Text* text;
     s32 ctx;
-    s32 num_players;
+    u8 num_players;
     s32 row_a;
     GXColor* color_ptr;
     GXColor* color2_ptr;
